+++ src/main/java/at/hannibal2/skyhanni/data/ItemClickData.kt
@@ -1,9 +1,13 @@
 package at.hannibal2.skyhanni.data
 
+import at.hannibal2.skyhanni.events.BlockClickEvent
 import at.hannibal2.skyhanni.events.EntityClickEvent
-import at.hannibal2.skyhanni.events.ItemClickInHandEvent
+import at.hannibal2.skyhanni.events.PacketEvent
 import at.hannibal2.skyhanni.utils.LorenzUtils
+import at.hannibal2.skyhanni.utils.toLorenzVec
 import net.minecraft.client.Minecraft
+import net.minecraft.network.play.client.C07PacketPlayerDigging
+import net.minecraft.network.play.client.C08PacketPlayerBlockPlacement
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import net.minecraftforge.fml.common.gameevent.InputEvent
 import org.lwjgl.input.Mouse
@@ -11,17 +15,17 @@ import org.lwjgl.input.Mouse
 class ItemClickData {
 
     @SubscribeEvent
-    fun onClick(event: InputEvent.MouseInputEvent) {
-        if (!Mouse.getEventButtonState()) return
-
-        val clickType = when (Mouse.getEventButton()) {
-            0 -> ClickType.LEFT_CLICK
-            1 -> ClickType.RIGHT_CLICK
-            else -> return
+    fun onItemClickSend(event: PacketEvent.SendEvent) {
+        val packet = event.packet
+        if (packet is C08PacketPlayerBlockPlacement) {
+            val position = packet.position.toLorenzVec()
+            BlockClickEvent(ClickType.RIGHT_CLICK, position, packet.stack).postAndCatch()
         }
-
-        val itemStack = Minecraft.getMinecraft().thePlayer.heldItem
-        ItemClickInHandEvent(clickType, itemStack).postAndCatch()
+        if (packet is C07PacketPlayerDigging && packet.status == C07PacketPlayerDigging.Action.START_DESTROY_BLOCK) {
+            val position = packet.position.toLorenzVec()
+            val itemInHand = Minecraft.getMinecraft().thePlayer.heldItem
+            BlockClickEvent(ClickType.LEFT_CLICK, position, itemInHand).postAndCatch()
+        }
     }
 
     @SubscribeEvent